@@ -3,12 +3,13 @@
   "version": "0.0.1",
   "private": true,
   "description": "A simple NextJS app with Plate",
+  "type": "module",
   "scripts": {
     "prebuild": "yarn build:contentlayer",
     "build": "yarn prebuild && yarn build:registry && next build",
-    "build:contentlayer": "contentlayer build",
-    "build:registry": "tsx --tsconfig ./scripts/tsconfig.scripts.json ./scripts/build-registry.ts",
-    "dev": "concurrently \"contentlayer dev\" \"NODE_OPTIONS=--max-old-space-size=16192 next dev\"",
+    "build:contentlayer": "contentlayer2 build",
+    "build:registry": "tsx --tsconfig ./scripts/tsconfig.scripts.json scripts/build-registry.mts",
+    "dev": "concurrently \"contentlayer2 dev\" \"NODE_OPTIONS=--max-old-space-size=16192 next dev\"",
     "dev:quick": "NODE_OPTIONS=--max-old-space-size=16192 next dev",
     "lint": "npx eslint . --ext .js,.jsx,.ts,.tsx",
     "lint:debug": "TIMING=1 DEBUG=eslint:cli-engine yarn lint:fix",
@@ -129,15 +130,14 @@
     "@vercel/analytics": "^1.2.2",
     "@vercel/og": "^0.6.2",
     "class-variance-authority": "^0.7.0",
-    "cmdk": "^0.2.1",
-    "contentlayer": "0.3.4",
+    "cmdk": "^0.2.0",
+    "contentlayer2": "^0.4.6",
     "date-fns": "^2.30.0",
     "framer-motion": "^10.18.0",
     "lodash.template": "^4.5.0",
-    "lucide-react": "^0.288.0",
-    "next": "14.1.4",
-    "next-axiom": "^1.1.1",
-    "next-contentlayer": "0.3.4",
+    "lucide-react": "^0.408.0",
+    "next": "14.3.0-canary.43",
+    "next-contentlayer2": "^0.4.6",
     "next-themes": "^0.2.1",
     "nuqs": "^1.17.1",
     "prism-react-renderer": "^1.3.5",
@@ -147,6 +147,7 @@
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.2.0",
     "react-lite-youtube-embed": "^2.4.0",
+    "react-resizable-panels": "^0.0.55",
     "react-syntax-highlighter": "^15.5.0",
     "react-tweet": "^3.2.0",
     "react-wrap-balancer": "^1.1.0",
@@ -157,16 +158,21 @@
     "slate-react": "0.102.0",
     "slate-test-utils": "1.3.2",
     "sonner": "^1.4.32",
+    "swr": "2.2.6-beta.3",
     "tailwind-merge": "^2.2.2",
+    "ts-morph": "^22.0.0",
     "unist-builder": "4.0.0",
-    "unist-util-visit": "^5.0.0"
+    "unist-util-visit": "^5.0.0",
+    "vaul": "^0.9.1"
   },
   "devDependencies": {
+    "@shikijs/compat": "^1.1.7",
     "@types/lodash.template": "^4.5.3",
     "@types/react-color": "^3.0.12",
     "@types/react-syntax-highlighter": "^15.5.11",
     "autoprefixer": "^10.4.18",
     "concurrently": "^8.2.2",
+    "glob": "^11.0.0",
     "mdast-util-toc": "^6.1.1",
     "postcss": "^8.4.38",
     "rehype": "^12.0.1",
@@ -175,14 +181,13 @@
     "rehype-slug": "^5.1.0",
     "remark": "^14.0.3",
     "remark-code-import": "^1.2.0",
-    "remark-gfm": "^3.0.1",
+    "remark-gfm": "^4.0.0",
     "rimraf": "^5.0.5",
-    "shiki": "^0.12.1",
+    "shiki": "^1.10.1",
     "tailwindcss": "^3.4.1",
     "tailwindcss-animate": "^1.0.7",
     "ts-node": "^10.9.2",
     "tsx": "^4.7.1",
-    "typescript": "5.4.2",
-    "vaul": "^0.9.0"
+    "typescript": "5.4.2"
   }
 }
