@@ -8,7 +8,7 @@ import { Index } from '@/__registry__';
 import { useConfig } from '@/hooks/use-config';
 import { styles } from '@/registry/styles';
 
-import { CopyButton, CopyWithClassNames } from './copy-button';
+import { CopyButton } from './copy-button';
 import { Icons } from './icons';
 import { StyleSwitcher } from './style-switcher';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from './ui/tabs';
@@ -16,8 +16,10 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from './ui/tabs';
 interface ComponentPreviewProps extends React.HTMLAttributes<HTMLDivElement> {
   name: string;
   align?: 'center' | 'end' | 'start';
+  description?: string;
   extractClassname?: boolean;
   extractedClassNames?: string;
+  hideCode?: boolean;
   padding?: 'md';
 }
 
@@ -25,8 +27,10 @@ export function ComponentPreview({
   align = 'start',
   children,
   className,
+  description,
   extractClassname,
   extractedClassNames,
+  hideCode = false,
   name,
   padding,
   ...props
@@ -57,7 +61,7 @@ export function ComponentPreview({
 
   const codeString = React.useMemo(() => {
     if (Code?.props['data-rehype-pretty-code-fragment'] !== undefined) {
-      const [, Button] = React.Children.toArray(
+      const [Button] = React.Children.toArray(
         Code.props.children
       ) as React.ReactElement[];
 
@@ -72,33 +76,45 @@ export function ComponentPreview({
     >
       <Tabs className="relative mr-auto w-full" defaultValue="preview">
         <div className="flex items-center justify-between pb-3">
-          <TabsList className="w-full justify-start rounded-none border-b bg-transparent p-0">
-            <TabsTrigger
-              className="relative h-9 rounded-none border-b-2 border-b-transparent bg-transparent px-4 pb-3 pt-2 font-semibold text-muted-foreground shadow-none transition-none data-[state=active]:border-b-primary data-[state=active]:text-foreground data-[state=active]:shadow-none"
-              value="preview"
-            >
-              Preview
-            </TabsTrigger>
-            <TabsTrigger
-              className="relative h-9 rounded-none border-b-2 border-b-transparent bg-transparent px-4 pb-3 pt-2 font-semibold text-muted-foreground shadow-none transition-none data-[state=active]:border-b-primary data-[state=active]:text-foreground data-[state=active]:shadow-none"
-              value="code"
-            >
-              Code
-            </TabsTrigger>
-          </TabsList>
+          {!hideCode && (
+            <TabsList className="w-full justify-start rounded-none border-b bg-transparent p-0">
+              <TabsTrigger
+                className="relative h-9 rounded-none border-b-2 border-b-transparent bg-transparent px-4 pb-3 pt-2 font-semibold text-muted-foreground shadow-none transition-none data-[state=active]:border-b-primary data-[state=active]:text-foreground data-[state=active]:shadow-none"
+                value="preview"
+              >
+                Preview
+              </TabsTrigger>
+              <TabsTrigger
+                className="relative h-9 rounded-none border-b-2 border-b-transparent bg-transparent px-4 pb-3 pt-2 font-semibold text-muted-foreground shadow-none transition-none data-[state=active]:border-b-primary data-[state=active]:text-foreground data-[state=active]:shadow-none"
+                value="code"
+              >
+                Code
+              </TabsTrigger>
+            </TabsList>
+          )}
         </div>
         <TabsContent className="relative rounded-md border" value="preview">
           {styles.length > 1 && (
             <div className="flex items-center justify-between p-4">
               <StyleSwitcher />
-              {extractedClassNames ? (
-                <CopyWithClassNames
-                  classNames={extractedClassNames}
+
+              <div className="flex items-center gap-2">
+                {/* {config.style === "default" && description ? ( */}
+                {/*  <V0Button */}
+                {/*    block={{ */}
+                {/*      code: codeString, */}
+                {/*      name, */}
+                {/*      style: config.style, */}
+                {/*      description, */}
+                {/*    }} */}
+                {/*  /> */}
+                {/* ) : null} */}
+                <CopyButton
+                  className="size-7 text-foreground opacity-100 hover:bg-muted hover:text-foreground [&_svg]:size-3.5"
                   value={codeString}
+                  variant="outline"
                 />
-              ) : (
-                codeString && <CopyButton value={codeString} />
-              )}
+              </div>
             </div>
           )}
           <React.Suspense
