@@ -2,7 +2,7 @@
 
 import * as React from 'react';
 
-import { allDocs } from 'contentlayer/generated';
+import { allDocs } from 'contentlayer2/generated';
 
 import { Mdx } from './mdx-components';
 
