@@ -10,7 +10,7 @@ function PageHeader({
   return (
     <section
       className={cn(
-        'flex flex-col items-start gap-2 px-4 pt-8 md:pt-12',
+        'mx-auto flex flex-col items-start gap-2 px-4 py-8 md:py-12 md:pb-8 lg:py-12 lg:pb-10',
         className
       )}
       {...props}
@@ -27,7 +27,7 @@ function PageHeaderHeading({
   return (
     <h1
       className={cn(
-        'text-3xl font-bold leading-tight tracking-tighter md:text-5xl lg:leading-[1.1]',
+        'text-3xl font-bold leading-tight tracking-tighter md:text-4xl lg:leading-[1.1]',
         className
       )}
       {...props}
@@ -42,7 +42,7 @@ function PageHeaderDescription({
   return (
     <Balance
       className={cn(
-        'max-w-[750px] text-lg text-muted-foreground sm:text-xl',
+        'max-w-2xl text-balance text-lg font-light text-foreground',
         className
       )}
       {...props}
@@ -50,4 +50,19 @@ function PageHeaderDescription({
   );
 }
 
-export { PageHeader, PageHeaderDescription, PageHeaderHeading };
+function PageActions({
+  className,
+  ...props
+}: React.HTMLAttributes<HTMLDivElement>) {
+  return (
+    <div
+      className={cn(
+        'flex w-full items-center justify-start gap-2 py-2',
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+export { PageActions, PageHeader, PageHeaderDescription, PageHeaderHeading };
