@@ -1,5 +1,5 @@
 import type { NavItem, NavItemWithChildren } from '@/types/nav';
-import type { Doc } from 'contentlayer/generated';
+import type { Doc } from 'contentlayer2/generated';
 
 import { cn } from '@udecode/cn';
 import Link from 'next/link';
@@ -45,6 +45,11 @@ export function DocsPager({ doc }: DocsPagerProps) {
 }
 
 export function getPagerForDoc(doc: Doc) {
+  // const nav = doc.slug.startsWith("/docs/charts")
+  //     ? docsConfig.chartsNav
+  //     : docsConfig.sidebarNav
+  //   const flattenedLinks = [null, ...flatten(nav), null]
+
   const flattenedLinks = [null, ...flatten(docsConfig.sidebarNav), null];
   const activeIndex = flattenedLinks.findIndex(
     (link) => doc.slug === link?.href
