@@ -17,7 +17,7 @@ import {
   SelectValue,
 } from './ui/select';
 
-export function StyleSwitcher({ className }: SelectTriggerProps) {
+export function StyleSwitcher({ className, ...props }: SelectTriggerProps) {
   const [config, setConfig] = useConfig();
 
   return (
@@ -32,6 +32,7 @@ export function StyleSwitcher({ className }: SelectTriggerProps) {
     >
       <SelectTrigger
         className={cn('h-7 w-[145px] text-xs [&_svg]:size-4', className)}
+        {...props}
       >
         <span className="text-muted-foreground">Style: </span>
         <SelectValue placeholder="Select style" />
